+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -324,6 +324,13 @@ public:
                     std::is_same_v<EvalMemorySpace, Kokkos::HostSpace>,
                     Kokkos::DefaultHostExecutionSpace,
                     Kokkos::DefaultExecutionSpace>;
+            // Check that BSplines are defined on the expected domain.
+            // The radial direction should extend to r=0
+            // The poloidal domain should have length 2*M_PI
+            assert(fabs(ddc::discrete_space<BSplinesR>().rmin()) < 1e-15);
+            assert(fabs(ddc::discrete_space<BSplinesTheta>().rmax()
+                        - ddc::discrete_space<BSplinesTheta>().rmin() - 2. * M_PI)
+                   < 1e-14);
             if constexpr (C > -1) {
                 IdxRange<BSplinesR> idx_range_r_ctrl_pts(IdxR(0), IdxStepR(std::max(2, C + 1)));
                 IdxRange<BSplinesR, BSplinesTheta> idx_range_ctrl_pts(
@@ -713,7 +720,7 @@ KOKKOS_FUNCTION Idx<BSplinesR, BSplinesTheta> PolarBSplines<BSplinesR, BSplinesT
             }
         }
         jmin_r = first_tensor_product_radial_spline;
-    } else {
+    } else if constexpr (n_singular_basis() > 0) {
         for (std::size_t k(0); k < n_singular_basis(); ++k) {
             singular_values(k) = 0.0;
         }