+++ toolchains/docker.gyselalibxx_env/environment.sh
@@ -10,5 +10,8 @@ fi
 current_folder=$(realpath $(dirname ${BASH_SOURCE[0]}))
 gyselalibxx_folder=$(realpath ${current_folder}/../..)
 
-docker run -v ${gyselalibxx_folder}:/src --workdir "/src" -it ghcr.io/gyselax/gyselalibxx_env
+cmake_prefixes="/opt/googletest:/opt/openmp/"
+
+docker pull ghcr.io/gyselax/gyselalibxx_env:latest
+docker run -v ${gyselalibxx_folder}:/src --workdir "/src" --user :$(id -g) -e CMAKE_PREFIX_PATH=${cmake_prefixes} -it ghcr.io/gyselax/gyselalibxx_env:latest