+++ toolchains/persee/v100/environment.sh
@@ -7,16 +7,14 @@ then
     exit 1
 fi
 
-if [ -f "$(which spack)" ]
+if command -v spack >/dev/null 2>&1
 then
-  spack env deactivate
+    spack env deactivate
 else
-  . /data/gyselarunner/spack-0.23.0/share/spack/setup-env.sh
+    . /data/gyselarunner/spack-1.1.0/share/spack/setup-env.sh
 fi
 
-spack load gcc@12
 spack env activate gyselalibxx-env-omp-cuda
 export OMP_PROC_BIND=spread
 export OMP_PLACES=threads
 export OMP_NUM_THREADS=8
-